+++ app/doc/[pdfId]/_components/index.ts
@@ -1,7 +1,7 @@
-import PDFChat from "./pdf-chat/PDFChat";
-import PDFNotes from "./pdf-notes/PDFNotes";
-import LoadingPDFNotes from "./pdf-notes/Loading";
-import PDFViewer from "./pdf-viewer/PDFViewer";
-import LoadingPDFViewer from "./pdf-viewer/Loading";
+import PDFChat from "./pdf-chat/pdf-chat";
+import PDFNotes from "./pdf-notes/pdf-notes";
+import LoadingPDFNotes from "./pdf-notes/loading";
+import PDFViewer from "./pdf-viewer/pdf-viewer";
+import LoadingPDFViewer from "./pdf-viewer/loading";
 
 export { PDFChat, PDFNotes, LoadingPDFNotes, PDFViewer, LoadingPDFViewer };
