+++ app/doc/[pdfId]/_components/pdf-chat/pdf-chat.tsx
