+++ app/doc/[pdfId]/_components/pdf-notes/pdf-notes.tsx
