+++ app/doc/[pdfId]/_components/pdf-viewer/pdf-viewer.tsx
